update str_to_int function to comply with new size_t type
As the new size_t type has been defined to stop invalid casting, I
realized that I hadn't deployed any size_t:s yet, so that's what I did.
This will help massively in reducing the ugliset type of code; casts.

+++ lib/conv.c
@@ -1,5 +1,6 @@
 #include "conv.h"
 #include "stdint.h"
+#include "stddef.h"
 
 void uint32_to_str(char* output, uint32_t number)
 {
@@ -93,7 +94,7 @@ char uint_to_char(uint number)
 }
 
 /* convert character array to integer */
-int str_to_int(char *array, int n)
+int str_to_int(char *array, size_t n)
 {    
     int number = 0;
     int mult = 1;
