+++ lib/conv.h
@@ -13,5 +13,5 @@ void int_to_str(char* output, int number);
 void short_to_str(char* output, short number);
 char uint32_to_char(uint32_t number);
 char uint_to_char(uint number);
-int str_to_int(char *array, int n);
+int str_to_int(char *array, size_t n);
 byte char_to_hex(char character);
